Move switch-matrix stuff to hardware module
