+++ src/main.rs
@@ -16,7 +16,6 @@ use hal::{
     Rtc, IO,
 };
 use hal::{prelude::*, Delay};
-use keyberon::debounce::Debouncer;
 use keyberon::key_code::KeyCode;
 use keyberon::layout::Layout;
 use usb_device::prelude::{UsbDeviceBuilder, UsbVidPid};
@@ -29,51 +28,10 @@ use usbd_human_interface_device::device::{
 use usbd_human_interface_device::page::Keyboard as HidKeyboard;
 use usbd_human_interface_device::prelude::*;
 
-mod board_modules;
-#[allow(non_upper_case_globals)]
-struct ButtonMatrix<'a, const InN: usize, const OutN: usize> {
-    ins: [&'a dyn InputPin<Error = Infallible>; InN],
-    outs: [&'a mut dyn OutputPin<Error = Infallible>; OutN],
-}
-#[allow(non_upper_case_globals)]
-struct BoardModule<'a, const InN: usize, const OutN: usize> {
-    matrix: ButtonMatrix<'a, InN, OutN>,
-    debouncer: Debouncer<[[bool; InN]; OutN]>,
-}
-#[allow(non_upper_case_globals)]
-impl<'a, const InN: usize, const OutN: usize> BoardModule<'a, InN, OutN> {
-    fn new(mut matrix: ButtonMatrix<'a, InN, OutN>, debounce_tolerance: u16) -> Self {
-        matrix.init();
-        Self {
-            matrix,
-            debouncer: ButtonMatrix::gen_debouncer(debounce_tolerance),
-        }
-    }
-}
+use crate::hardware::matrix::{ButtonMatrix, BoardModule};
 
-#[allow(non_upper_case_globals)]
-impl<'a, const InN: usize, const OutN: usize> ButtonMatrix<'a, InN, OutN> {
-    fn init(&mut self) {
-        for out in &mut self.outs {
-            let _ = out.set_high();
-        }
-    }
-    fn key_scan(&mut self, delay: &mut Delay) -> [[bool; InN]; OutN] {
-        let mut res = [[false; InN]; OutN];
-        for (out_dx, out_pin) in self.outs.iter_mut().enumerate() {
-            let _todo_logerr = out_pin.set_low();
-            delay.delay_us(5u32);
-            for (in_dx, in_pin) in self.ins.iter().enumerate() {
-                res[out_dx][in_dx] = in_pin.is_low().unwrap();
-            }
-            let _todo_logerr = out_pin.set_high();
-        }
-        res
-    }
-    fn gen_debouncer(n: u16) -> Debouncer<[[bool; InN]; OutN]> {
-        Debouncer::new([[false; InN]; OutN], [[false; InN]; OutN], n)
-    }
-}
+mod board_modules;
+mod hardware;
 
 static mut USB_MEM: [u32; 1024] = [0; 1024];
 
@@ -151,6 +109,7 @@ fn main() -> ! {
         .product("totally not a malicious thing")
         .device_class(3)
         .build();
+
 
     let left_finger = ButtonMatrix {
         ins: [
